Completed stringify
Edited master function (deals with objects) to include a switch that
skips over undefined values and functions. Had not come across
'continue' before. Got it to work but very unclear on the behaviour of
switches. Was under the impressions default only occurred if other
switch options failed, however, it now seems to occur unless 'continue'
is specified. Confuddled.